+++ components/reports/plugin_formatters/stdout/healthmap.rb
@@ -20,7 +20,7 @@ class PluginFormatters::HealthMap < Arachni::Plugin::Formatter
             state = i.keys[0]
             url   = i.values[0]
 
-            if state == :unsafe
+            if state == :with_issues
                 print_bad( url )
             else
                 print_ok( url )
@@ -30,8 +30,8 @@ class PluginFormatters::HealthMap < Arachni::Plugin::Formatter
         print_line
 
         print_info "Total: #{results[:total]}"
-        print_ok "Without issues: #{results[:safe]}"
-        print_bad "With issues: #{results[:unsafe]} ( #{results[:issue_percentage].to_s}% )"
+        print_ok "Without issues: #{results[:without_issues]}"
+        print_bad "With issues: #{results[:with_issues]} ( #{results[:issue_percentage].to_s}% )"
     end
 
 end